+++ graylog2-server/src/main/java/org/graylog2/web/IndexHtmlGenerator.java
@@ -16,61 +16,17 @@
  */
 package org.graylog2.web;
 
-import com.floreysoft.jmte.Engine;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.io.Resources;
-import org.graylog2.configuration.HttpConfiguration;
-import org.graylog2.rest.RestTools;
-
-import javax.inject.Inject;
-import javax.inject.Singleton;
 import javax.ws.rs.core.MultivaluedMap;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
-import java.util.Map;
 
-import static java.util.Objects.requireNonNull;
-
-@Singleton
-public class IndexHtmlGenerator {
-    private final String template;
-    private final List<String> cssFiles;
-    private final List<String> sortedJsFiles;
-    private final Engine templateEngine;
-    private final HttpConfiguration httpConfiguration;
-
-    @Inject
-    public IndexHtmlGenerator(final PluginAssets pluginAssets,
-                              final Engine templateEngine,
-                              final HttpConfiguration httpConfiguration) throws IOException {
-        this(
-                Resources.toString(Resources.getResource("web-interface/index.html.template"), StandardCharsets.UTF_8),
-                pluginAssets.cssFiles(),
-                pluginAssets.sortedJsFiles(),
-                templateEngine,
-                httpConfiguration);
-    }
-
-    private IndexHtmlGenerator(final String template,
-                               final List<String> cssFiles,
-                               final List<String> sortedJsFiles,
-                               final Engine templateEngine,
-                               final HttpConfiguration httpConfiguration) throws IOException {
-        this.template = requireNonNull(template, "template");
-        this.cssFiles = requireNonNull(cssFiles, "cssFiles");
-        this.sortedJsFiles = requireNonNull(sortedJsFiles, "sortedJsFiles");
-        this.templateEngine = requireNonNull(templateEngine, "templateEngine");
-        this.httpConfiguration = requireNonNull(httpConfiguration, "httpConfiguration");
-    }
-
-    public String get(MultivaluedMap<String, String> headers) {
-        final Map<String, Object> model = ImmutableMap.<String, Object>builder()
-                .put("title", "Graylog Web Interface")
-                .put("cssFiles", cssFiles)
-                .put("jsFiles", sortedJsFiles)
-                .put("appPrefix", RestTools.buildExternalUri(headers, httpConfiguration.getHttpExternalUri()))
-                .build();
-        return templateEngine.transform(template, model);
-    }
+/**
+ * Implementations provide HTML content for an "index.html" file. This file will be served to browser clients.
+ */
+public interface IndexHtmlGenerator {
+    /**
+     * Get the HTML content.
+     *
+     * @param headers the HTTP request headers of the web request
+     * @return the HTML string
+     */
+    String get(MultivaluedMap<String, String> headers);
 }
